Parse response.body instead of response
Before if there was a the response code was a 200 or 201 the response was being parsed.
This causes an error when creating a new user because nothing is returned from BambooHR
except in the headers.  By parsing response.body instead, there is no error becuase response.body
defaults to an empty string even is nothing is returned from BambooHR.