Fix Bamboozled::API::Base#request
Prior to this change I was trying to parse the successful response returned from the BambooHR API after creating a new employee with JSON.parse.  This, however, threw an error becuase reponse.body was an empty string and JSON.parse can't handle an empty string, throwing the error: 'JSON::ParserError: A JSON text must at least contain two octets!'. To solve this problem I'm checking if response is nil. If it is return the headers so the end user can get the newly created employees location (url) and BambooHR ID in case they need it, if not parse the response as before.