Change Bamboozled::API::Base#request to check if response.body is empty
We need to check if the response.body is empty or not to determine what to return from the method.
When a successful POST request is made response.body returns an empty string and when a successful GET
request is made it returns JSON or XML. Note the use of response.body.to_s before calling empty? becuase
could not figure out how to have Webmock's stub_request method return an empty string for the response.body