Improve response handinling on job_info
The tests imply that the response from the API would be xml, but in the
case of `client.example.job_info`, the returned data is an Array in JSON
format.

When we call `with_indifferent_access` on the array, we get a method
missing exception, which causes the code to then try and parse a JSON
response as xml.

This change only calls the method when we get an object back.