Remove Active Support dependency
This commit removes **Active Support** as a dependency in favor of defaulting to
strings for `Hash` access. Active Support is currently being used solely for a
couple of instances of `hash_with_indifferent_access`. We can keep `Bamboozled`
smaller and delegate the use of such libraries to the user when reading a
response.

Resolves: #23