Fix Issue #36: client.meta.users failing
The issue stems from the various ways the Bamboo HR API returns results
for the metadata endpoints. The original code tried to handle it with
metaprogramming, but the returning results are not the same.

Added tests and methods specific to the four metadata endpoints.

Address [Issue #36](https://github.com/Skookum/bamboozled/issues/36)